feat/dev-back: Database integer fields
-- Database --
* Changed amount, roi and duration type from INT to BIGINT, because it supports big integer.

+++ back/database/entrypoint.sql
@@ -21,13 +21,13 @@ USE `investdb` ;
 CREATE TABLE IF NOT EXISTS `investdb`.`financials` (
   `id` INT NOT NULL AUTO_INCREMENT,
   `name` VARCHAR(255) NOT NULL,
-  `amount` INT NOT NULL,
+  `amount` BIGINT NOT NULL,
   `date` DATE NOT NULL,
   `description` VARCHAR(255) NOT NULL,
   `status` VARCHAR(255) NOT NULL,
-  `roi` INT NOT NULL,
+  `roi` BIGINT NOT NULL,
   `location` VARCHAR(255) NOT NULL,
-  `duration` INT NOT NULL,
+  `duration` BIGINT NOT NULL,
   PRIMARY KEY (`id`),
   UNIQUE INDEX `name_UNIQUE` (`name` ASC) VISIBLE)
 ENGINE = InnoDB;
@@ -38,13 +38,13 @@ ENGINE = InnoDB;
 CREATE TABLE IF NOT EXISTS `investdb`.`estates` (
   `id` INT NOT NULL AUTO_INCREMENT,
   `name` VARCHAR(255) NOT NULL,
-  `amount` INT NOT NULL,
+  `amount` BIGINT NOT NULL,
   `date` DATE NOT NULL,
   `description` VARCHAR(255) NOT NULL,
   `status` VARCHAR(255) NOT NULL,
-  `roi` INT NOT NULL,
+  `roi` BIGINT NOT NULL,
   `location` VARCHAR(255) NOT NULL,
-  `duration` INT NOT NULL,
+  `duration` BIGINT NOT NULL,
   PRIMARY KEY (`id`),
   UNIQUE INDEX `name_UNIQUE` (`name` ASC) VISIBLE)
 ENGINE = InnoDB;
